loop checking really works now

+++ csdl_alpha/src/operations/loops/loop.py
@@ -3,7 +3,7 @@ from csdl_alpha.src.graph.variable import Variable
 
 class IterationVariable(Variable):
     def __init__(self, vals):
-        super().__init__(vals[0])
+        super().__init__(value=vals[0])
         self.shape = (1,)
         self.vals = vals
         self.name = 'iter'
@@ -11,7 +11,6 @@ class IterationVariable(Variable):
 class LoopVariable(Variable):
     def __init__(self, var1, var2):
         super().__init__(shape=(var1.shape)) #NOTE: idk if we even need to do this
-        self.shape=None
         self.var1 = var1
         self.var2 = var2
 
@@ -34,7 +33,8 @@ class Loop(Operation):
         for output in self.outputs:
             self.recorder.active_graph.add_node(output)
 
-    def compute_inline(self):
+    def compute_inline(self, *args):
+        # print(self.graph)
         for i in range(len(self.vals)):
             self.iter_var.set_value(self.vals[i])
             for loop_var in self.loop_vars:
@@ -43,12 +43,14 @@ class Loop(Operation):
                 else:
                     loop_var.set_value(loop_var.var2.value)
             self.graph.execute_inline()
+        return [output.value for output in self.outputs]
 
 
 class vrange():
     def __init__(self, lower:int=0, upper:int = 100, increment:int=1, *, vals:list[int] = None, check:bool=False):
         
         self.check = check
+        self.disallow_feedback = True # turning off feedback for now
 
         # process runtime iterations
         if vals is None:
@@ -61,21 +63,27 @@ class vrange():
             self.vals = vals
 
         self.curr_index = 0
-        self.max_index = len(self.vals)
+        if self.check:
+            self.max_index = len(self.vals)
+        else:
+            self.max_index = 2
 
         # enter new graph
         from csdl_alpha.api import manager
         self._recorder = manager.active_recorder
         self._recorder._enter_subgraph(add_missing_variables=True)
         self._graph = self._recorder.active_graph
+        self._graph_node = self._recorder.active_graph_node
 
         # initialize iteration variable:
         self.iteration_variable = IterationVariable(self.vals)
 
-    def get_ops_and_shapes(self):
+    def get_ops_and_shapes(self, graph=None):
         ops = []
         shapes = []
-        for node in self._graph.node_table.keys():
+        if graph is None:
+            graph = self._recorder.active_graph
+        for node in graph.node_table.keys():
             if isinstance(node, Operation):
                 ops.append(type(node))
             elif isinstance(node, Variable):
@@ -103,6 +111,9 @@ class vrange():
         for input in self.iter1_inputs:
             self.iter1_non_inputs.discard(input)
 
+        # don't want iteration variable to be removed, even if it's not used
+        self.iter1_non_inputs.discard(self.iteration_variable)
+
         # deleting the operations so we cana find inputs to the second iteration in the same way
         self._graph._delete_nodes(ops)
 
@@ -123,15 +134,19 @@ class vrange():
             if not input1 is input2: 
                 if input2 in self.iter1_outputs:
                     # this implies input 1 comes from outside the loop and input 2 comes from the first iteration
-                    # we want to go from input2 to the corresponding output of the 2nd iteration
-                    output2 = self.iter2_outputs[self.iter1_outputs.index(input2)]
-                    loop_var = LoopVariable(input1, output2)
-                    loop_vars.append(loop_var)
-                    self._graph._delete_nodes([input1])
-                    self.iter1_non_inputs.remove(input2)
-                    self._graph._replace_node(input2, loop_var)
+                    if self.disallow_feedback:
+                        raise ValueError(f'Feedback from variable {input1} found')
+                    else:
+                        # we want to go from input2 to the corresponding output of the 2nd iteration
+                        output2 = self.iter2_outputs[self.iter1_outputs.index(input2)]
+                        loop_var = LoopVariable(input1, output2)
+                        loop_vars.append(loop_var)
+                        self._graph._delete_nodes([input1])
+                        self.iter1_non_inputs.remove(input2)
+                        self._graph._replace_node(input2, loop_var)
                 else:
                     # this implies input 1 and input 2 are both made in the loop, so we can just keep input 2
+                    self._graph._delete_nodes([input1])
                     pass
 
         # delete any remnanats of the first iteration
@@ -145,9 +160,20 @@ class vrange():
 
         # add the loop operation to the graph
         #NOTE: this only exposes outputs of operations, not variables created within the loop
-        op = Loop(external_inputs, self.iter2_outputs, self._graph, self.vals, self.iteration_variable, loop_vars)
-        if self._recorder.inline:
-            op.compute_inline()
+        self.op = Loop(
+            list(external_inputs), 
+            self.iter2_outputs, 
+            self._graph, 
+            self.vals, 
+            self.iteration_variable, 
+            loop_vars
+            )
+
+    def _check_ops_and_shapes(self, ops, shapes):
+        if ops != self.ops:
+            raise ValueError(f'Operations changed between iterations')
+        if shapes != self.shapes:
+            raise ValueError(f'Shapes changed between iterations')
 
 
     def __next__(self):
@@ -158,42 +184,35 @@ class vrange():
             if self.check:
                 self.ops, self.shapes = self.get_ops_and_shapes()
             self.post_iteration_one()
-        # second iteration - check ops and shapes havent changed, find feedback
-        elif self.curr_index == 2:
+            if self.check:
+                self._recorder._enter_subgraph(add_missing_variables=True)
+                self._recorder._add_node(self.iteration_variable)
+        # last iteration - check feedback, check ops and shapes, end loop
+        elif self.curr_index == self.max_index:
+            final = True
             self.post_iteration_two()
             if self.check:
+                # need to pass in graph here because we exited it in post_iteration_two
+                ops, shapes = self.get_ops_and_shapes(self._graph)
+                self._check_ops_and_shapes(ops, shapes)
+        # second to last iteration - need to re-enter main graph
+        elif self.curr_index == self.max_index-1:
+            if self.check:
                 ops, shapes = self.get_ops_and_shapes()
-                if ops != self.ops:
-                    raise ValueError(f'Operations changed between iterations')
-                if shapes != self.shapes:
-                    raise ValueError(f'Shapes changed between iterations')
-            else:
-                final=True
-        # final iteration - check ops and shapes, end loop
-        elif self.curr_index >= self.max_index:
-            self._recorder._enter_subgraph()
-            graph = self._recorder.active_graph
-            ops, shapes = self.get_ops_and_shapes()
+                self._check_ops_and_shapes(ops, shapes)
+                self._recorder._delete_current_graph()
+        # all other iterations - enter/exit dummy graph, check ops and shapes
+        elif self.curr_index > 1:
             if self.check:
-                if ops != self.ops:
-                    raise ValueError(f'Operations changed between iterations')
-                if shapes != self.shapes:
-                    raise ValueError(f'Shapes changed between iterations')
-            self._recorder._delete_current_graph()
-            final = True
-        # all other iterations - check ops and shapes
-        elif self.curr_index > 2:
-            self._recorder._enter_subgraph()
-            graph = self._recorder.active_graph
-            ops, shapes = self.get_ops_and_shapes()
-            if self.check:
-                if ops != self.ops:
-                    raise ValueError(f'Operations changed between iterations')
-                if shapes != self.shapes:
-                    raise ValueError(f'Shapes changed between iterations')
-            self._recorder._delete_current_graph()
+                ops, shapes = self.get_ops_and_shapes()
+                self._check_ops_and_shapes(ops, shapes)
+                self._recorder._delete_current_graph()
+                self._recorder._enter_subgraph(add_missing_variables=True)
+                self._recorder._add_node(self.iteration_variable)
 
         if final:
+            if self._recorder.inline:
+                self.op.compute_inline()
             raise StopIteration
 
         self.curr_index+=1
@@ -210,14 +229,55 @@ if __name__ == '__main__':
     recorder.start()
     a = csdl.Variable(value=2, name='a')
     b = csdl.Variable(value=3, name='b')
-    j = 0
-    for i in vrange(0, 10, check=True):
-        print(j)
-        b = a + b
-        c = a*2
-        j+=1
+    k = 0
 
-    print(b.value) # should be 23
+    # for i in vrange(0, 10, check=True):
+    #     b2 = a + b
+    #     c = a*2
+
+    # print(b2.value)
+    # print(c.value)
+
+    # print(recorder.active_graph)
+
+    # top_graph_node = recorder.active_graph_node
+    # outer_loop_graph_node = top_graph_node.children[0]
+
+    # top_graph = top_graph_node.value
+    # outer_loop_graph = outer_loop_graph_node.value
+
+    # top_graph.visualize('top_graph')
+    # outer_loop_graph.visualize('outer_loop_graph')
+
+
+    for i in vrange(0, 10, check=True):
+        print(f'begin outer iteration {k}')
+        k += 1
+        l = 0
+        for j in vrange(0, 10, check=True):
+            print(f'inner iteration {l}')
+            l += 1
+            d = i*2
+            e = i*j
+            b2 = a + b
+            c = a*2
+        print(f'end outer iteration {k}')
+
+    top_graph_node = recorder.active_graph_node
+    outer_loop_graph_node = top_graph_node.children[0]
+    inner_loop_graph_node = outer_loop_graph_node.children[0]
+
+    top_graph = top_graph_node.value
+    outer_loop_graph = outer_loop_graph_node.value
+    inner_loop_graph = inner_loop_graph_node.value
+
+    top_graph.visualize('top_graph')
+    outer_loop_graph.visualize('outer_loop_graph')
+    inner_loop_graph.visualize('inner_loop_graph')
+
+    print(d.value) # should be 18
+    print(e.value) # should be 81
+    print(b2.value) # should be 5
     print(c.value) # should be 4
     recorder.active_graph.visualize('outer_graph')
     recorder.stop()
